pyproject.toml
[project]
name = "bambulab-mqtt"
version = "1.0.1"
description = "A simple Python MQTT client for Bambu Lab printers"
authors = [
    { name = "William" }
]
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }

dependencies = [
    "paho-mqtt>=1.6.1"
]

[project.urls]
Homepage = "https://github.com/Willgob/Bambulab-MQTT-python-library"
Repository = "https://github.com/Willgob/Bambulab-MQTT-python-library"
